Tugas 8

+++ app/Http/Controllers/BukusController.php
@@ -5,13 +5,28 @@ namespace App\Http\Controllers;
 use App\Models\Buku;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Validator;
 
 
 class BukusController extends Controller {
 
     public function index(Request $request){
-        $bukus = Buku::where(['id' => Auth::user()->id])->OrderBy("id", "DESC")->paginate(2)->toArray();
+
+        if(Gate::denies('read-buku')) {
+            return response()->json([
+                'success' => false,
+                'status' => 403,
+                'message' => 'You are unauthorized'
+            ], 403);
+        }
+
+        if(Auth::user()->role === 'admin') {
+            $bukus = Buku::OrderBy("id", "DESC")->paginate(2)->toArray();
+        }else {
+            $bukus = Buku::where(['id' => Auth::user()->id])->OrderBy("id", "DESC")->paginate(2)->toArray();
+        }
+
         $response = [
             "total_count" => $bukus["total"],
             "limit" => $bukus["per_page"],
@@ -26,6 +41,19 @@ class BukusController extends Controller {
     
     public function store(Request $request){
         $input = $request->all();
+        $buku = Buku::create($input);
+
+        // Authtorization
+
+        if(Gate::denies('create-buku', $buku)) {
+            return response()->json([
+                'success' => false,
+                'status' => 403,
+                'message' => 'You are unauthorized',
+            ]);
+        }
+        // authorization end
+
         $validationRules = [
             // 'buku_id' => 'required|exists:users,id',
             'buku_id' => 'required|min:3',
@@ -53,6 +81,17 @@ class BukusController extends Controller {
             abort(404);
         }
 
+        // Authtorization
+
+        if(Gate::denies('read-detail-buku', $buku)) {
+            return response()->json([
+                'success' => false,
+                'status' => 403,
+                'message' => 'You are unauthorized',
+            ]);
+        }
+        // authorization end
+
         return response()->json($buku, 200);
     }
 
@@ -63,6 +102,18 @@ class BukusController extends Controller {
         if(!$buku){
             abort(404);
         }
+
+        // Authtorization
+        // check if current user is authorized to do this action
+
+        if(Gate::denies('update-buku', $buku)) {
+            return response()->json([
+                'success' => false,
+                'status' => 403,
+                'message' => 'You are unauthorized',
+            ]);
+        }
+        // authorization end
 
         $validationRules = [
             'buku_id' => 'required|min:3',
@@ -92,10 +143,21 @@ class BukusController extends Controller {
         if(!$buku){
             abort(404);
         }
+
+        // Authtorization
+
+        if(Gate::denies('delete-buku', $buku)) {
+            return response()->json([
+                'success' => false,
+                'status' => 403,
+                'message' => 'You are unauthorized',
+            ]);
+        }
+        // authorization end
         
         $buku->delete();
-        $message = ['message' => 'deleted successfully', 'buku_id'];
+        $message = ['message' => 'deleted successfully', 'id' => $id];
         
-        return response()->json($buku, 200);
+        return response()->json($message, 200);
     }
 }
